+++ src/Contracts/CacheableContract.php
@@ -1,0 +1,74 @@
+<?php
+
+/*
+ * NOTICE OF LICENSE
+ *
+ * Part of the Rinvex Repository Package.
+ *
+ * This source file is subject to The MIT License (MIT)
+ * that is bundled with this package in the LICENSE file.
+ *
+ * Package: Rinvex Repository Package
+ * License: The MIT License (MIT)
+ * Link:    https://rinvex.com
+ */
+
+namespace Rinvex\Repository\Contracts;
+
+interface CacheableContract
+{
+    /**
+     * Set the repository cache lifetime.
+     *
+     * @param float|int $cacheLifetime
+     *
+     * @return $this
+     */
+    public function setCacheLifetime($cacheLifetime);
+
+    /**
+     * Get the repository cache lifetime.
+     *
+     * @return float|int
+     */
+    public function getCacheLifetime();
+
+    /**
+     * Set the repository cache driver.
+     *
+     * @param string $cacheDriver
+     *
+     * @return $this
+     */
+    public function setCacheDriver($cacheDriver);
+
+    /**
+     * Get the repository cache driver.
+     *
+     * @return string
+     */
+    public function getCacheDriver();
+
+    /**
+     * Enable repository cache clear.
+     *
+     * @param bool $status
+     *
+     * @return $this
+     */
+    public function enableCacheClear($status);
+
+    /**
+     * Determine if repository cache clear is enabled.
+     *
+     * @return bool
+     */
+    public function isCacheClearEnabled();
+
+    /**
+     * Forget the repository cache.
+     *
+     * @return $this
+     */
+    public function forgetCache();
+}