+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 
 
+## [v2.0.1] - 2016-08-06
+- Add missing argument to simplePaginate method (#54)
+- Add config option for default model directory (close #55)
+- Extract Cache methods to a Contract / Trait (close #57)
+- Fix spelling typos & fix docs
+- Added extra logic to paginate and simplePaginate methods, fixes #61 (#62)
+
 ## [v2.0.0] - 2016-07-01
 - Drop `findOrCreate` method (close #33)
 - Change `retrieveModel` behavior (close #34)
@@ -56,6 +63,7 @@ This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 ## v1.0.0 - 2016-06-18
 - Commit first draft
 
+[v2.0.1]: https://github.com/rinvex/repository/compare/v2.0.0...v2.0.1
 [v2.0.0]: https://github.com/rinvex/repository/compare/v1.0.5...v2.0.0
 [v1.0.5]: https://github.com/rinvex/repository/compare/v1.0.4...v1.0.5
 [v1.0.4]: https://github.com/rinvex/repository/compare/v1.0.3...v1.0.4